update tests and fix C interop
Returning false could fail since the size
of a bool is different between C, C++ and C#.
This is fixed by adding wrappers which assure that
a byte is returned containing 1 for true and 0 for false.
Furthermore this commit adds missing attributes to ref
parameters in NativeRcl or turns them into out parameters.